move test files to prevent bundling
