+++ vscode/npm-package/jest.config.cjs
@@ -3,7 +3,7 @@ const common = require('./jest.common.config.cjs')
 /** @returns {Promise<import('jest').Config>} */
 module.exports = async () => {
   // When debugging, we want to have a longer timeout
-  let testTimeout = 5000; 
+  let testTimeout = 100000; 
   if (process.env.VSCODE_INSPECTOR_OPTIONS) {
       testTimeout = 999999;
   }
