+++ vscode/npm-package/jest.config.integration.cjs
@@ -2,9 +2,9 @@ const common = require('./jest.common.config.cjs')
 
 module.exports = {
   ...common,
-  globalSetup: "<rootDir>/integration_tests/setup.ts",
-  globalTeardown: "<rootDir>/integration_tests/teardown.ts",
+  globalSetup: "<rootDir>/tests/integration/setup.ts",
+  globalTeardown: "<rootDir>/tests/integration/teardown.ts",
   testMatch: [
-    '**/integration_tests/?(*.)+(spec).ts?(x)'
+    '**/tests/integration/?(*.)+(spec).ts?(x)'
   ]
 };
