+++ vscode/npm-package/tests/integration/setup.ts
@@ -1,5 +1,5 @@
-import { ensureKiotaIsPresent } from '../install';
 import { Config } from '@jest/types';
+import { ensureKiotaIsPresent } from '../../install';
 
 export default async (globalConfig: Config.GlobalConfig, projectConfig: Config.ProjectConfig) => {
     // Ensure kiota binary to optimize integration tests
