+++ vscode/npm-package/tests/integration/teardown.ts
