+++ vscode/npm-package/tests/unit/generatePlugin.spec.ts
@@ -1,6 +1,6 @@
-import { generatePlugin } from '../lib/generatePlugin';
-import { KiotaLogEntry } from '../types';
-import { KiotaPluginType } from '../types';
+import { generatePlugin } from '../../lib/generatePlugin';
+import { KiotaLogEntry } from '../../types';
+import { KiotaPluginType } from '../../types';
 import { setupKiotaStubs } from './stubs.util';
 
 describe("generate plugin", () => {
