+++ vscode/npm-package/tests/unit/getKiotaTree.spec.ts
@@ -1,5 +1,5 @@
-import { KiotaTreeResult } from '../types';
-import { getKiotaTree } from '../lib/getKiotaTree';
+import { KiotaTreeResult } from '../../types';
+import { getKiotaTree } from '../../lib/getKiotaTree';
 import { setupKiotaStubs } from './stubs.util';
 
 describe("get kiota tree", () => {
