+++ vscode/npm-package/tests/unit/getManifestDetails.spec.ts
@@ -1,5 +1,5 @@
-import { KiotaManifestResult } from '../types';
-import { getManifestDetails } from '../lib/getManifestDetails';
+import { KiotaManifestResult } from '../../types';
+import { getManifestDetails } from '../../lib/getManifestDetails';
 import { setupKiotaStubs } from './stubs.util';
 
 describe("get manifest details", () => {
