+++ vscode/npm-package/tests/unit/languageInformation.spec.ts
@@ -1,4 +1,4 @@
-import { getLanguageInformationForDescription, getLanguageInformationInternal } from '../lib/languageInformation';
+import { getLanguageInformationForDescription, getLanguageInformationInternal } from '../../lib/languageInformation';
 import { setupKiotaStubs } from './stubs.util';
 
 const sampleLanguageInfo = {
